requirements.txt
langchain
langchain-community
langchain-cohere
fastapi
uvicorn 
python-dotenv 
pypdf 
pandas
